requirements.txt
flask==2.3.3  
werkzeug==2.3.7  
flask-sqlalchemy==3.1.1  
flask-login==0.6.2  
flask-bcrypt==1.0.1  
python-dotenv==1.0.0  
flask-wtf==1.2.1  
email-validator==2.1.0